java-seq 0.75

+++ xss/idioms/java-seq/libs/IntInterpolator.java
@@ -7,8 +7,8 @@ public class IntInterpolator implements IInterpolator
 
 	public Object interpolate(Object from, Object to, double t) 
 	{
-		Double v1 = (Double)from;
-		Double v2 = (Double)to;
+		Integer v1 = (Integer)from;
+		Integer v2 = (Integer)to;
 		return (Integer)((int)(v1 + (v2 - v1)*t));
 	}
 }
